Move helper to shared file, add change to other controllers
