+++ modules/decision_reviews/app/controllers/decision_reviews/v1/supplemental_claims_controller.rb
@@ -173,16 +173,6 @@ module DecisionReviews
         req_body_obj
       end
 
-      def handle_area_code_for_schema_validation(req_body_obj)
-        phone = req_body_obj.dig('data', 'attributes', 'veteran', 'phone')
-        area_code = phone&.dig('areaCode')
-
-        return req_body_obj if area_code.is_a?(String) && !area_code.empty?
-
-        req_body_obj['data']['attributes']['veteran']['phone']&.delete('areaCode')
-        req_body_obj
-      end
-
       def merge_evidence_entries(entries)
         merged_entry = entries.first.deep_dup
         merged_attributes = merged_entry['attributes']
